+++ src/utils/object.js
@@ -1,5 +1,9 @@
-export const keys = Object.keys;
+export const { keys } = Object;
 
 export function blank () {
 	return Object.create( null );
+}
+
+export function forOwn ( object, func ) {
+	Object.keys( object ).forEach( key => func( object[ key ], key ) );
 }