+++ test/function/skips-dead-branches-g/main.js
@@ -1,0 +1,6 @@
+var a = 0;
+var b = 1;
+var c = true ? a : b;
+var d = false ? a : b;
+
+console.log( c + d );